+++ tests/testthat/tests.make_stancode.R
@@ -1938,7 +1938,7 @@ test_that("argument 'stanvars' is handled correctly", {
   expect_match2(scode, "vector[K] M;")
   expect_match2(scode, "matrix[K, K] V;")
   
-  # define a hierachical prior on the regression coefficients
+  # define a hierarchical prior on the regression coefficients
   bprior <- set_prior("normal(0, tau)", class = "b") +
     set_prior("target += normal_lpdf(tau | 0, 10)", check = FALSE)
   stanvars <- stanvar(scode = "real<lower=0> tau;", 
@@ -1947,6 +1947,22 @@ test_that("argument 'stanvars' is handled correctly", {
                          prior = bprior, stanvars = stanvars)
   expect_match2(scode, "real<lower=0> tau;")
   expect_match2(scode, "target += normal_lpdf(b | 0, tau);")
+  
+  # ensure that variables are passed to the likelihood of a threaded model
+  foo <- 0.5
+  stanvars <- stanvar(foo) +
+    stanvar(scode = "real<lower=0> tau;", 
+            block = "parameters", pll_args = "real tau")
+    
+  scode <- make_stancode(count ~ 1, epilepsy, family = poisson(),
+                         stanvars = stanvars, threads = threading(2),
+                         parse = FALSE)
+  expect_match2(scode, 
+    "partial_log_lik_lpmf(int[] seq, int start, int end, int[] Y, real Intercept, real foo, real tau)"
+  )
+  expect_match2(scode, 
+    "reduce_sum(partial_log_lik_lpmf, seq, grainsize, Y, Intercept, foo, tau)"
+  )
   
   # add transformation at the end of a block
   stanvars <- stanvar(scode = "  r_1_1 = r_1_1 * 2;",